# apps/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_domain_coupon_handler.CreateStoreCouponRequest:
    properties:
      description:
        type: string
      max_per_user:
        type: integer
      price:
        type: number
      status:
        type: string
      terms:
        type: string
      title:
        type: string
      total_quantity:
        type: integer
      type:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    required:
    - max_per_user
    - title
    - total_quantity
    - type
    type: object
  internal_domain_coupon_handler.InitiatePaymentRequest:
    properties:
      userId:
        type: string
    required:
    - userId
    type: object
  internal_domain_coupon_handler.ValidateCouponRequest:
    properties:
      quantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for RevieU
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RevieU Core API
  version: "1.0"
paths:
...
  /coupons/{id}/payment/initiate:
    post:
      consumes:
      - application/json
      description: Initiates payment flow for a coupon
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: integer
      - description: Initiate payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_coupon_handler.InitiatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate coupon payment
      tags:
      - coupon
  /coupons/{id}/redeem:
    post:
      description: Redeems a coupon for the authenticated user
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redeem coupon
      tags:
      - coupon
  /coupons/{id}/validate:
    post:
      consumes:
      - application/json
      description: Validates a coupon by ID
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: integer
      - description: Validate coupon request
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_domain_coupon_handler.ValidateCouponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate coupon
      tags:
      - coupon
...
  /merchant/stores/{id}/coupons:
    post:
      consumes:
      - application/json
      description: Creates a store-scoped coupon for an owned published store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create store coupon request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_coupon_handler.CreateStoreCouponRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create store coupon
      tags:
      - coupon
  /merchant/stores/{id}/coupons/{couponId}:
    delete:
      description: Soft-deletes a store-scoped coupon under an owned store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Coupon ID
        in: path
        name: couponId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete store coupon
      tags:
      - coupon
...
  /packages/{id}:
    get:
      description: Returns a package by ID
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get package detail
      tags:
      - package
...
  /stores/{id}/coupons:
    get:
      description: Lists published active coupons under a store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List store coupons
      tags:
      - coupon
...
  /user/reviews:
    get:
      description: Returns reviews created by the authenticated user
      parameters:
      - description: Cursor
        in: query
        name: cursor
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List my reviews
      tags:
      - content
...
  /users/{id}/reviews:
    get:
      description: Returns a user's reviews
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List user's reviews
      tags:
      - content
...
schemes:
- http
- https
swagger: "2.0"